+++ src/TestableHttpClient/TestableHttpMessageHandlerOptions.cs
@@ -1,4 +1,6 @@
-﻿namespace TestableHttpClient;
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace TestableHttpClient;
 
 public sealed class TestableHttpMessageHandlerOptions
 {
@@ -8,5 +10,8 @@ public sealed class TestableHttpMessageHandlerOptions
         PropertyNamingPolicy = JsonNamingPolicy.CamelCase
     };
 
-    public RoutingOptions RoutingOptions { get; } = new RoutingOptions();
+    [ExcludeFromCodeCoverage]
+    [Obsolete("Renamed to UriPatternMatchingOptions")]
+    public RoutingOptions RoutingOptions => UriPatternMatchingOptions;
+    public UriPatternMatchingOptions UriPatternMatchingOptions { get; } = new UriPatternMatchingOptions();
 }