+++ src/TestableHttpClient/Utils/RoutingResponseBuilder.cs
@@ -7,5 +7,5 @@ internal class RoutingResponseBuilder : IRoutingResponseBuilder
     public RoutingResponse RoutingResponse { get; } = new();
 
     public void MapFallBackResponse(IResponse fallBackResponse) => RoutingResponse.FallBackResponse = fallBackResponse;
-    public void Map(string route, IResponse response) => RoutingResponse.ResponseMap.Add(RouteParser.Parse(route), response);
+    public void Map(string route, IResponse response) => RoutingResponse.ResponseMap.Add(UriPatternParser.Parse(route), response);
 }