+++ src/TestableHttpClient/Utils/UriPatternParserException.cs
@@ -3,35 +3,35 @@
 namespace TestableHttpClient.Utils;
 
 /// <summary>
-/// Exception thrown when the route string can not be parsed correctly.
+/// Exception thrown when the URI pattern string can not be parsed correctly.
 /// </summary>
 [SuppressMessage("SonarSource", "S3925", Justification = "These exceptions don't need to be serialized.")]
-public sealed class RouteParserException : Exception
+public sealed class UriPatternParserException : Exception
 {
     /// <summary>
-    /// Initializes a new instance of the <see cref="RouteParserException"/> class with the default error message.
+    /// Initializes a new instance of the <see cref="UriPatternParserException"/> class with the default error message.
     /// </summary>
-    public RouteParserException()
+    public UriPatternParserException()
         : this("Route parsing failed.")
     {
     }
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="RouteParserException"/> class with a specified error message.
+    /// Initializes a new instance of the <see cref="UriPatternParserException"/> class with a specified error message.
     /// </summary>
     /// <param name="message">The error message that explains the reason for the exception.</param>
-    public RouteParserException(string message) : base(message)
+    public UriPatternParserException(string message) : base(message)
     {
     }
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="RouteParserException"/>
+    /// Initializes a new instance of the <see cref="UriPatternParserException"/>
     /// class with a specified error message and a reference to the inner exception that is the cause of
     /// this exception.
     /// </summary>
     /// <param name="message">The error message that explains the reason for the exception.</param>
     /// <param name="innerException">The exception that is the cause of the current exception.</param>
-    public RouteParserException(string message, Exception innerException) : base(message, innerException)
+    public UriPatternParserException(string message, Exception innerException) : base(message, innerException)
     {
     }
 }