+++ test/TestableHttpClient.IntegrationTests/AssertingRequests.cs
@@ -95,10 +95,9 @@ public class AssertingRequests
 
         _ = await client.GetAsync("https://httpbin.org/get?email=admin@example.com");
 
-        testHandler.ShouldHaveMadeRequests().WithQueryString("email=admin@example.com");
-        testHandler.ShouldHaveMadeRequests().WithQueryString("email=*");
-        Assert.Throws<HttpRequestMessageAssertionException>(() => testHandler.ShouldHaveMadeRequests().WithQueryString(""));
-        Assert.Throws<HttpRequestMessageAssertionException>(() => testHandler.ShouldHaveMadeRequests().WithQueryString("email=admin%40example.com"));
+        testHandler.ShouldHaveMadeRequests().WithRequestUri("?email=admin@example.com");
+        testHandler.ShouldHaveMadeRequests().WithRequestUri("?email=*");
+        Assert.Throws<HttpRequestMessageAssertionException>(() => testHandler.ShouldHaveMadeRequests().WithRequestUri("?email=admin%40example.com"));
     }
 
     [Fact]